# backend/services/ai-chatbot/openapi.yaml
swagger: '2.0'
info:
  title: Grassroots Football Management API
  description: API Gateway for Grassroots Football Management microservices
  version: 1.0.0
host: gfm-1ykrsotxghm8m.apigateway.grassroots-football-management.cloud.goog
...
schemes:
  - https
securityDefinitions:
  firebase:
    type: oauth2
    flow: "implicit"
    authorizationUrl: ""
    x-google-issuer: "https://securetoken.google.com/grassroots-football-management"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "grassroots-football-management"
security:
  - firebase: []
paths:
...
  /schedule/fixture:
    post:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Add a new fixture to the schedule.
      operationId: addFixture
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [homeTeam, awayTeam, ageGroup, division, date, createdBy]
            properties:
              homeTeam:
                type: string
              awayTeam:
                type: string
              ageGroup:
                type: string
              division:
                type: string
              date:
                type: string
                format: date-time
              createdBy:
                type: string
      responses:
        '201':
          description: Fixture added successfully.
        '500':
          description: Internal server error.

    get:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Get fixtures for a month.
      operationId: getFixturesByMonth
      parameters:
        - in: query
          name: month
          required: true
          type: string
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
      responses:
        '200':
          description: List of fixtures.
          schema:
            type: array
            items:
              type: object
        '500':
          description: Internal server error.

    put:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Update fixture details.
      operationId: updateFixture
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [matchId, clubName, ageGroup, division]
            properties:
              matchId:
                type: string
              homeTeam:
                type: string
              awayTeam:
                type: string
              date:
                type: string
                format: date-time
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
      responses:
        '200':
          description: Fixture updated successfully.
        '500':
          description: Internal server error.

    delete:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Delete a fixture.
      operationId: deleteFixture
      parameters:
        - in: query
          name: matchId
          required: true
          type: string
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
      responses:
        '200':
          description: Fixture deleted successfully.
        '500':
          description: Internal server error.
  /schedule/fixture/{matchId}:
    get:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app

        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Get fixture by match ID.
      operationId: getFixtureById
      parameters:
        - in: path
          name: matchId
          required: true
          type: string
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
      responses:
        '200':
          description: Fixture details.
        '404':
          description: Fixture not found.
        '500':
          description: Internal server error.
  /schedule/fixtures:
    get:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Get all fixtures for age group and division.
      operationId: getAllFixtures
      parameters:
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
      responses:
        '200':
          description: List of all fixtures.
        '500':
          description: Internal server error.
  /schedule/training:
    post:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Add a training session.
      operationId: addTraining
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [date, location, createdBy]
            properties:
              date:
                type: string
                format: date-time
              location:
                type: string
              notes:
                type: string
              createdBy:
                type: string
      responses:
        '201':
          description: Training added successfully.
        '500':
          description: Internal server error.

    get:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Get trainings for a month.
      operationId: getTrainingsByMonth
      parameters:
        - in: query
          name: month
          required: true
          type: string
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
      responses:
        '200':
          description: List of trainings.
          schema:
            type: array
            items:
              type: object
        '500':
          description: Internal server error.

    put:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Update training details.
      operationId: updateTraining
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [trainingId]
            properties:
              trainingId:
                type: string
              date:
                type: string
                format: date-time
              location:
                type: string
              notes:
                type: string
      responses:
        '200':
          description: Training updated successfully.
        '500':
          description: Internal server error.

    delete:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Delete a training session.
      operationId: deleteTraining
      parameters:
        - in: query
          name: trainingId
          required: true
          type: string
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
      responses:
        '200':
          description: Training deleted successfully.
        '500':
          description: Internal server error.
  /schedule/trainings:
    get:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Get all trainings.
      operationId: getAllTrainings
      parameters:
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
      responses:
        '200':
          description: List of trainings.
          schema:
            type: array
            items:
              type: object
        '500':
          description: Internal server error.
  /schedule/training/availability:
    post:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Update training availability
      description: Updates the availability status of a user for a specific training session
      operationId: updateTrainingAvailability
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - email
              - clubName
              - ageGroup
              - division
              - available
              - trainingId
            properties:
              email:
                type: string
              clubName:
                type: string
              ageGroup:
                type: string
              division:
                type: string
              available:
                type: boolean
              trainingId:
                type: string
      responses:
        '200':
          description: Availability updated
        '500':
          description: Internal server error

    get:
      x-google-backend:
        address: https://microservice-schedule-management-180320540339.europe-west2.run.app
        jwtAudience: https://microservice-schedule-management-180320540339.europe-west2.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      summary: Get training availability
      description: Fetches availability of all players for a specific training session
      operationId: getTrainingAvailability
      produces:
        - application/json
      parameters:
        - in: query
          name: clubName
          required: true
          type: string
        - in: query
          name: ageGroup
          required: true
          type: string
        - in: query
          name: division
          required: true
          type: string
        - in: query
          name: trainingId
          required: true
          type: string
      responses:
        '200':
          description: List of availability responses
        '400':
          description: Missing or invalid parameters
        '500':
          description: Internal server error
...